Rename zendesk-setup to zendesk_setup
Following ruby conventions
