+++ lib/get-user-numbers-2012.rb
@@ -1,6 +1,6 @@
 require 'zendesk_api'
 
-require_relative 'zendesk-setup.rb'
+require_relative 'zendesk_setup'
 
 #
 # 2012
